requirements.txt
facenet-pytorch
torch
torchvision
scikit-learn
numpy
pillow
flask
google-generativeai
